feat(baking): support handlers whose resolved type is not the same as persisted model
