+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/resource/NamedImageHandler.kt
@@ -11,8 +11,8 @@ import com.netflix.spinnaker.keel.api.ec2.NamedImage
 import com.netflix.spinnaker.keel.clouddriver.CloudDriverService
 import com.netflix.spinnaker.keel.events.TaskRef
 import com.netflix.spinnaker.keel.persistence.ResourceRepository
+import com.netflix.spinnaker.keel.plugin.ResourceDiff
 import com.netflix.spinnaker.keel.plugin.ResourceHandler
-import com.netflix.spinnaker.keel.plugin.ResourceHandler.ResourceDiff
 import com.netflix.spinnaker.keel.plugin.ResourceNormalizer
 import kotlinx.coroutines.runBlocking
 import org.slf4j.Logger
