+++ keel-plugin/src/main/kotlin/com/netflix/spinnaker/keel/plugin/ResourceDiff.kt
@@ -1,0 +1,10 @@
+package com.netflix.spinnaker.keel.plugin
+
+import de.danielbechler.diff.node.DiffNode
+
+data class ResourceDiff<T : Any>(
+  val source: T,
+  val diff: DiffNode
+) {
+  constructor(source: T) : this(source, DiffNode.newRootNode())
+}
