+++ keel-plugin/src/main/kotlin/com/netflix/spinnaker/keel/plugin/ResourceHandler.kt
@@ -15,154 +15,15 @@
  */
 package com.netflix.spinnaker.keel.plugin
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import com.netflix.spinnaker.keel.api.ApiVersion
 import com.netflix.spinnaker.keel.api.Resource
-import com.netflix.spinnaker.keel.api.ResourceKind
-import com.netflix.spinnaker.keel.api.ResourceMetadata
-import com.netflix.spinnaker.keel.api.ResourceName
-import com.netflix.spinnaker.keel.api.SubmittedResource
-import com.netflix.spinnaker.keel.api.randomUID
-import com.netflix.spinnaker.keel.events.TaskRef
-import com.netflix.spinnaker.keel.exceptions.InvalidResourceStructureException
-import de.danielbechler.diff.node.DiffNode
-import org.slf4j.Logger
 
-interface ResourceHandler<T : Any> : KeelPlugin {
-  data class ResourceDiff<T : Any>(
-    val source: T,
-    val diff: DiffNode
-  ) {
-    constructor(source: T) : this(source, DiffNode.newRootNode())
-  }
-
-  val log: Logger
-
-  val apiVersion: ApiVersion
+interface ResourceHandler<T : Any> : ResolvableResourceHandler<T, T> {
 
   /**
-   * Maps the kind to the implementation type.
+   * This method is a no-op when the model and resolved types are the same.
    */
-  val supportedKind: Pair<ResourceKind, Class<T>>
-
-  val objectMapper: ObjectMapper
-
-  val normalizers: List<ResourceNormalizer<*>>
-
-  /**
-   * Validates the resource spec, and generates a metadata header, and applies any defaults /
-   * normalization.
-   *
-   * Implementors should not generally need to override this method,
-   *  custom normalization and validation should be done with a [ResourceNormalizer].
-   *
-   * @return a hydrated `Resource` with a name generated by convention, a `uid`, default values
-   * applied, etc.
-   */
-  fun normalize(resource: SubmittedResource<Any>): Resource<T> {
-    val spec: T
-    try {
-      spec = objectMapper.convertValue(resource.spec, supportedKind.second)
-    } catch (e: IllegalArgumentException) {
-      throw InvalidResourceStructureException(
-        "Submitted resource with an incorrect structure, cannot convert to ${resource.kind} ${resource.spec}",
-        resource.spec.toString(),
-        e
-      )
-    }
-    val metadata = ResourceMetadata(
-      name = generateName(spec),
-      resourceVersion = 0L,
-      uid = randomUID()
-    )
-    val hydratedResource = Resource(
-      apiVersion = resource.apiVersion,
-      kind = resource.kind,
-      metadata = metadata,
-      spec = spec
-    )
-    @Suppress("UNCHECKED_CAST")
-    return normalize(hydratedResource as Resource<Any>)
-  }
-
-  /**
-   * Generate a unique name for a resource based on its spec.
-   */
-  fun generateName(spec: T): ResourceName
-
-  /**
-   * Validates the resource and applies any defaults / normalization.
-   *
-   * @return [resource] or a copy of [resource] that may have been changed, for example in order to
-   * set default values.
-   */
-  @Suppress("UNCHECKED_CAST")
-  fun normalize(resource: Resource<Any>): Resource<T> {
-    var normalizedResource = resource.copy(
-      spec = objectMapper.convertValue(resource.spec, supportedKind.second)
-    ) as Resource<T>
-    for (normalizer in normalizers) {
-      if (normalizer.handles(resource.apiVersion, resource.kind)) {
-        log.debug("Normalizing ${resource.metadata.name} with ${normalizer.javaClass}")
-        normalizedResource = normalizer.normalize(normalizedResource) as Resource<T>
-      }
-    }
-    return normalizedResource
-  }
-
-  /**
-   * Return the current _actual_ representation of what [resource] looks like in the cloud.
-   * The entire desired state is passed so that implementations can use whatever identifying
-   * information they need to look up the resource. Implementations of this method should not
-   * actuate any changes.
-   */
-  fun current(resource: Resource<T>): T?
-
-  /**
-   * Create a resource so that it matches the desired state represented by [resource].
-   *
-   * By default this just delegates to [upsert].
-   *
-   * Implement this method and [update] if you need to handle create and update in different ways.
-   * Otherwise just implement [upsert].
-   *
-   * @return a list of tasks launched to actuate the resource.
-   */
-  fun create(resource: Resource<T>): List<TaskRef> =
-    upsert(resource, null)
-
-  /**
-   * Update a resource so that it matches the desired state represented by [resource].
-   *
-   * By default this just delegates to [upsert].
-   *
-   * Implement this method and [create] if you need to handle create and update in different ways.
-   * Otherwise just implement [upsert].
-   *
-   * @return a list of tasks launched to actuate the resource.
-   */
-  fun update(
-    resource: Resource<T>,
-    resourceDiff: ResourceDiff<T> = ResourceDiff(resource.spec)
-  ): List<TaskRef> =
-    upsert(resource, resourceDiff)
-
-  /**
-   * Create or update a resource so that it matches the desired state represented by [resource].
-   *
-   * You don't need to implement this method if you are implementing [create] and [update]
-   * individually.
-   *
-   * @return a list of tasks launched to actuate the resource.
-   */
-  fun upsert(resource: Resource<T>, resourceDiff: ResourceDiff<T>? = null): List<TaskRef> {
-    TODO("Not implemented")
-  }
-
-  /**
-   * Delete a resource as the desired state is that it should no longer exist.
-   */
-  fun delete(resource: Resource<T>)
+  override fun desired(resource: Resource<T>): T = resource.spec
 }
 
 /**
